remove mpd.conf from dotfiles altogether
I use MPD as a service *as well as* a user program. Because of that, I
configure it like a service -- therefore mpd.conf shouldn't be .mpdconf
and shouldn't really be in dotfiles at all.
mpd.conf is now in my adminfiles, where it belongs. If you're in an
environment where you can't edit /etc files (e.g. not a desktop
computer), clone my adminfiles anyway and edit it ¯\_(ツ)_/¯

Signed-off-by: Ben Orchard <thefirstmuffinman@gmail.com>